Remove the deprecated opcodes and deal with the fallout
