+++ ex3_shared_libs/common/src/lib.rs
@@ -1,3 +1,6 @@
+pub mod component_ids;
+pub use component_ids::{ComponentIds};
+
 /// Ports used for communication between handlers and simulated subsystems / payloads
 pub mod ports {
     pub const SIM_DFGM_PORT: u16 = 1802;
@@ -14,110 +17,6 @@ pub mod ports {
     pub const BULK_MSG_HANDLER_DISPATCHER_PORT: u16 = 1903;
 }
 
-/// Each thing that can emit or receive a message has an associated ID. Each message header includes this id for source and destination.
-/// Referencing this page:
-pub mod component_ids {
-    use std::fmt::{self};
-    use std::str::FromStr;
-    use strum::EnumIter;
-
-    // ---------- Depricated but left to not break things -------- //
-    pub const OBC: u8 = 0;
-    pub const EPS: u8 = 1;
-    pub const ADCS: u8 = 2;
-    pub const DFGM: u8 = 3;
-    pub const IRIS: u8 = 4;
-    pub const GPS: u8 = 5;
-    //.....
-    pub const GS: u8 = 7;
-    pub const COMS: u8 = 8;
-    // ----------------------------------------------------------- //
-
-    #[derive(EnumIter, PartialEq, Debug)]
-    pub enum ComponentIds {
-        OBC = 0,
-        EPS = 1,
-        ADCS = 2,
-        DFGM = 3,
-        IRIS = 4,
-        GPS = 5,
-        DEPLOYABLES = 6,
-        GS = 7,
-        COMS = 8,
-        BulkMsgDispatcher = 9,
-        SHELL = 10,
-        UHF = 11,
-        LAST = 12,
-    }
-
-    impl fmt::Display for ComponentIds {
-        fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-            match *self {
-                ComponentIds::OBC => write!(f, "OBC"),
-                ComponentIds::EPS => write!(f, "EPS"),
-                ComponentIds::ADCS => write!(f, "ADCS"),
-                ComponentIds::DFGM => write!(f, "DFGM"),
-                ComponentIds::IRIS => write!(f, "IRIS"),
-                ComponentIds::GPS => write!(f, "GPS"),
-                ComponentIds::DEPLOYABLES => write!(f, "DEPLOYABLES"),
-                ComponentIds::GS => write!(f, "GS"),
-                ComponentIds::COMS => write!(f, "COMS"),
-                ComponentIds::BulkMsgDispatcher => write!(f, "BulkMsgDispatcher"),
-                ComponentIds::SHELL => write!(f, "SHELL"),
-                ComponentIds::UHF => write!(f, "UHF"),
-                ComponentIds::LAST => write!(f, "illegal"),
-            }
-        }
-    }
-    impl FromStr for ComponentIds {
-        type Err = ();
-        fn from_str(s: &str) -> Result<Self, Self::Err> {
-            match s {
-                "OBC" => Ok(ComponentIds::OBC),
-                "EPS" => Ok(ComponentIds::EPS),
-                "ADCS" => Ok(ComponentIds::ADCS),
-                "DFGM" => Ok(ComponentIds::DFGM),
-                "IRIS" => Ok(ComponentIds::IRIS),
-                "GPS" => Ok(ComponentIds::GPS),
-                "DEPLOYABLES" => Ok(ComponentIds::DEPLOYABLES),
-                "GS" => Ok(ComponentIds::GS),
-                "COMS" => Ok(ComponentIds::COMS),
-                "BulkMsgDispatcher" => Ok(ComponentIds::BulkMsgDispatcher),
-                "SHELL" => Ok(ComponentIds::SHELL),
-                "UHF" => Ok(ComponentIds::UHF),
-                "LAST" => Err(()),
-                _ => Err(()),
-            }
-        }
-    }
-
-    //TODO - Find a way to make this return a result type instead of panicking
-    //       - the 'From<u8> method from std::convert lib does not allow for returning a Result type
-    impl TryFrom<u8> for ComponentIds {
-        type Error = ();
-
-        fn try_from(value: u8) -> Result<Self, Self::Error> {
-            match value {
-                x if x == ComponentIds::OBC as u8 => Ok(ComponentIds::OBC),
-                x if x == ComponentIds::EPS as u8 => Ok(ComponentIds::EPS),
-                x if x == ComponentIds::ADCS as u8 => Ok(ComponentIds::ADCS),
-                x if x == ComponentIds::DFGM as u8 => Ok(ComponentIds::DFGM),
-                x if x == ComponentIds::IRIS as u8 => Ok(ComponentIds::IRIS),
-                x if x == ComponentIds::GPS as u8 => Ok(ComponentIds::GPS),
-                x if x == ComponentIds::DEPLOYABLES as u8 => Ok(ComponentIds::DEPLOYABLES),
-                x if x == ComponentIds::GS as u8 => Ok(ComponentIds::GS),
-                x if x == ComponentIds::COMS as u8 => Ok(ComponentIds::COMS),
-                x if x == ComponentIds::BulkMsgDispatcher as u8 => {
-                    Ok(ComponentIds::BulkMsgDispatcher)
-                }
-                x if x == ComponentIds::SHELL as u8 => Ok(ComponentIds::SHELL),
-                x if x == ComponentIds::UHF as u8 => Ok(ComponentIds::UHF),
-                x if x == ComponentIds::LAST as u8 => Err(()),
-                _ => Err(()),
-            }
-        }
-    }
-}
 
 /// For constants that are used across the entire project
 pub mod constants {
